refactor: move middleware options to separate file
